# api/collaboration/openapi.yaml
openapi: 3.1.0
info:
  title: AphelionDMM Collaboration Control API
  version: 1.0.0
  description: Version 1 lifecycle and snapshot contract. Filesystem paths and process commands are never accepted.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://127.0.0.1
  - url: https://collaboration.example
paths:
...
  /v1/hosted/sessions:
    post:
      summary: Create a durable hosted collaboration session
      operationId: createHostedSession
      security: [{ bearerToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSessionRequest' }
      responses:
        '201':
          description: Durable hosted session created with the authenticated identity as owner
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HostedSessionResponse' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '413': { $ref: '#/components/responses/Error' }
...
  /v1/sessions/{session_id}/hosted-invitations:
    parameters: [{ $ref: '#/components/parameters/SessionID' }]
    post:
      summary: Create a one-use hosted invitation
      operationId: createHostedInvitation
      security: [{ bearerToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateHostedInvitationRequest' }
      responses:
        '201':
          description: One-use hosted invitation. The raw token is returned only in this response.
          headers:
            Cache-Control: { schema: { type: string, const: no-store } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HostedInvitation' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/sessions/{session_id}/hosted-invitations/redeem:
    parameters: [{ $ref: '#/components/parameters/SessionID' }]
    post:
      summary: Redeem a one-use hosted invitation
      operationId: redeemHostedInvitation
      security: [{ bearerToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RedeemHostedInvitationRequest' }
      responses:
        '200':
          description: Hosted membership created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HostedMembership' }
        '401': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    launchToken: { type: http, scheme: bearer, description: Single-use embedded launch token }
    bearerToken: { type: http, scheme: bearer }
  parameters:
    SessionID:
      name: session_id
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
  responses:
    Error:
      description: Bounded stable error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    SHA256:
      type: string
      pattern: '^[0-9a-f]{64}$'
...
    CreateSessionRequest:
      type: object
      additionalProperties: false
      required: [snapshot]
      properties:
        snapshot: { $ref: '#/components/schemas/Snapshot' }
        display_name: { type: string, minLength: 1, maxLength: 128, default: Owner }
...
    HostedSessionResponse:
      type: object
      additionalProperties: false
      required: [session_id, document_id, revision, map_hash]
      properties:
        session_id: { type: string, minLength: 1, maxLength: 128 }
        document_id: { type: string, format: uuid }
        revision: { type: integer, minimum: 0 }
        map_hash: { $ref: '#/components/schemas/SHA256' }
...
    CreateHostedInvitationRequest:
      type: object
      additionalProperties: false
      required: [role]
      properties:
        role: { type: string, enum: [viewer, editor] }
    HostedInvitation:
      type: object
      additionalProperties: false
      required: [token, role, expires_at]
      properties:
        token: { type: string, minLength: 1, maxLength: 128 }
        role: { type: string, enum: [viewer, editor] }
        expires_at: { type: string, format: date-time }
    RedeemHostedInvitationRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token: { type: string, minLength: 1, maxLength: 128 }
    HostedMembership:
      type: object
      additionalProperties: false
      required: [session_id, actor_id, role]
      properties:
        session_id: { type: string, minLength: 1, maxLength: 128 }
        actor_id: { type: string, format: uuid }
        role: { type: string, enum: [viewer, editor] }
...
    Snapshot:
      type: object
      additionalProperties: false
      required: [protocol_version, schema_version, document_id, revision, environment_hash, max_x, max_y, max_z, tiles]
      properties:
        protocol_version: { type: integer, const: 1 }
        schema_version: { type: integer, const: 1 }
        document_id: { type: string, format: uuid }
        revision: { type: integer, minimum: 0 }
        environment_hash: { $ref: '#/components/schemas/SHA256' }
        max_x: { type: integer, minimum: 1 }
        max_y: { type: integer, minimum: 1 }
        max_z: { type: integer, minimum: 1 }
        tiles: { type: array, maxItems: 16777216, items: { type: object } }
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: { type: string, minLength: 1, maxLength: 128 }
        message: { type: string, minLength: 1, maxLength: 512 }
        current_revision: { type: integer, minimum: 0 }
        current_map_hash: { $ref: '#/components/schemas/SHA256' }
...